added react-bootstrap, moved to pages folder, and started some code for the login/sign up
